Restructure the server into callflow
